requirements.txt
mkdocs-material
google-generativeai
flask
flask-cors
